requirements.txt
requests
schedule
boto3
python-dotenv
